Remove CommentsContainer. Move comments for post to route. Add console error
