--- src/components/CommentsContainer/index.tsx
@@ -1,191 +1,0 @@
-import * as React from 'react';
-import { IPost } from '../../types/reddit';
-import { MdKeyboardBackspace } from 'react-icons/md';
-import Comment from '../Comment';
-import { Group, Paper, Stack, Box, Title, Center, Avatar, useMantineTheme } from '@mantine/core';
-import { useRouter } from 'next/router';
-import { trpc } from '../../utils/trpc';
-import Post from '../Post';
-import { useMediaQuery } from '@mantine/hooks';
-import { useQueryClient } from 'react-query';
-import { useSelector } from 'react-redux';
-import { postSelector, PostsState } from 'src/redux/slices';
-import SortSelect from '../MobileSelect/SortSelect';
-import { Comments, IStory, Prompt } from 'src/types/db';
-import VirtualizedDataDisplay from '../VirtualizedDataDisplay';
-import AccountDrawer from '../AccountDrawer';
-import { useUser } from 'src/hooks/useUser';
-import { useStyles } from './comments.styles';
-
-export const CollapseContext = React.createContext<{
-  state: Record<string, boolean>;
-  collapseComment: (id: string, value?: boolean) => void;
-}>({
-  state: {},
-  collapseComment: function (id: string, value?: boolean): void {
-    this.state[id] = value ?? !this.state[id];
-  },
-});
-export const useCollapsedState = () => React.useContext(CollapseContext);
-
-interface Props {
-  post?: IPost;
-  postId: string;
-}
-const CommentsContainer = ({ postId }: Props) => {
-  const theme = useMantineTheme();
-  const largeScreen = useMediaQuery('(min-width: 900px)');
-
-  const router = useRouter();
-  const [drawerOpen, setDrawerOpen] = React.useState(false);
-
-  const user = useUser();
-  const queryClient = useQueryClient();
-
-  const postInfo = useSelector((state: PostsState) => postSelector(state, postId));
-
-  const storiesDownloaded = useSelector((state: PostsState) => state.stories[postId]);
-
-  const [collapsedState, setCollapsedState] = React.useState<Record<string, boolean>>({});
-
-  const {
-    data: storiesData,
-    isLoading,
-    isFetching,
-    isRefetching,
-    isError,
-    error,
-  } = trpc.useQuery(['story.forPost', { id: postId, userId: user.userId }], {
-    initialData: () => {
-      if (storiesDownloaded === undefined) {
-        return;
-      }
-      return storiesDownloaded;
-    },
-  });
-  const { data: postData } = trpc.useQuery(['post.byId', { id: postId, userId: user.userId }], {
-    initialData: () => {
-      if (postInfo === undefined) {
-        const queryCache = queryClient.getQueryData(['post.sort']) as Prompt[];
-        if (queryCache !== undefined) {
-          const cacheInfo = queryCache.find((val) => val.id === postId);
-          if (cacheInfo) {
-            // console.log('Cached post: ', cacheInfo);
-            return cacheInfo;
-          }
-        }
-        return;
-      }
-      const { ...rest } = postInfo;
-      return rest;
-    },
-  });
-  const { classes } = useStyles({ largeScreen });
-
-  const noComments = postData?.totalComments === 0;
-
-  const getParentComments = (comments: Comments) => {
-    return Object.values(comments).filter((val) => val.mainCommentId === null);
-  };
-
-  const collapseComment = React.useCallback(
-    (id: string, value?: boolean) => {
-      setCollapsedState((p) => ({ ...p, [id]: value ?? !p[id] }));
-    },
-    [postId]
-  );
-
-  React.useEffect(() => {
-    if (storiesData) {
-      setCollapsedState(
-        Object.values(storiesData).reduce((acc, v) => {
-          acc[v.id] = false;
-          return acc;
-        }, {} as Record<string, boolean>)
-      );
-    }
-  }, [storiesData]);
-
-  return (
-    <Stack align="center">
-      <Stack spacing={0} sx={{ width: '100%' }}>
-        <Paper px="lg" py="xs" className={classes.header}>
-          <Group noWrap align="start" position="apart" sx={{ width: '100%' }}>
-            <MdKeyboardBackspace
-              size={30}
-              onClick={() => {
-                router.back();
-              }}
-            />
-            <Avatar
-              radius={'xl'}
-              onClick={() => {
-                setDrawerOpen(true);
-              }}
-            />
-            <AccountDrawer opened={drawerOpen} closeDrawer={() => setDrawerOpen(false)} />
-          </Group>
-        </Paper>
-        {/* Post Details */}
-        {postData && (
-          <Box mt={60}>
-            <Post
-              {...postData}
-              totalComments={postData.totalComments}
-              isDownloaded={postInfo?.downloaded}
-              liked={postData.liked}
-            />
-          </Box>
-        )}
-        <Stack spacing={0} pb={40}>
-          <Group
-            noWrap
-            px="lg"
-            py="xs"
-            sx={{
-              backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[5] : theme.colors.gray[1],
-            }}
-          >
-            <SortSelect />
-          </Group>
-          {noComments ? (
-            <Center sx={{ height: '50vh' }}>
-              <Title order={2} color={theme.colors.dark[3]}>
-                No Stories Yet
-              </Title>
-            </Center>
-          ) : (
-            <CollapseContext.Provider value={{ state: collapsedState, collapseComment }}>
-              <VirtualizedDataDisplay
-                dataInfo={{
-                  error,
-                  isError,
-                  isFetching,
-                  isLoading,
-                  isRefetching,
-                  data: storiesData ? getParentComments(storiesData) : [],
-                }}
-                renderItem={(item: IStory & { replies: string[] }) => {
-                  return (
-                    <Comment
-                      key={item.id}
-                      {...item}
-                      allReplies={storiesData ?? {}}
-                      replies={item.replies}
-                      postId={postId}
-                      postAuthor={postData?.author ?? ''}
-                      replyIndex={0}
-                      isDownloaded={postInfo?.downloaded}
-                    />
-                  );
-                }}
-              />
-            </CollapseContext.Provider>
-          )}
-        </Stack>
-      </Stack>
-    </Stack>
-  );
-};
-
-export default CommentsContainer;
